re-organize RecoMuon/L3TrackFinder such that all plugins are defined in plugins folder
