+++ RecoMuon/L3TrackFinder/plugins/EtaPhiEstimator.h
